Moved Utils tests and added tests for Overloaded
