+++ kernel/UnitTests.cpp
@@ -16,6 +16,7 @@
 #include "UnitTests/Framework.h"
 #include "UnitTests/MemoryManagerTests.h"
 #include "UnitTests/PrintTests.h"
+#include "UnitTests/UtilsTests.h"
 #include "MemoryManager.h"
 #include "Print.h"
 #include "Utils.h"
@@ -26,46 +27,6 @@
 namespace
 {
     using namespace ::UnitTests;
-
-    ///////////////////////////////////////////////////////////////////////////
-    // Utils.h tests
-    ///////////////////////////////////////////////////////////////////////////
-
-    // #TODO: no obvious way to test MemoryMappedIO functions at this time
-    // #TODO: no obvious way to test Timing functions at this time
-
-    /**
-     * Ensure WriteMultiBitValue and ReadMultiBitValue set and read the expected bits
-    */
-    void ReadWriteMultiBitValueTest()
-    {
-        static constexpr uint64_t inputValue = 0xABCD; // larger than the mask to ensure we're masking
-        static constexpr uint64_t mask = 0xFF;
-        static constexpr uint64_t shift = 3;
-
-        std::bitset<64> bitset;
-        WriteMultiBitValue(bitset, inputValue, mask, shift);
-        EmitTestResult(bitset.to_ullong() == 0x0000'0000'0000'0668, "Write multi bit value mask/shift");
-        EmitTestResult(ReadMultiBitValue<uint64_t>(bitset, mask, shift) == 0xCD, "Read/write multi bit value round-trip");
-    }
-
-    /**
-     * Ensure WriteMultiBitValue and ReadMultiBitValue work with enum values
-    */
-    void ReadWriteMultiBitEnumTest()
-    {
-        enum class TestEnum: uint8_t
-        {
-            Value = 0b101
-        };
-        static constexpr uint64_t mask = 0b111;
-        static constexpr uint64_t shift = 3;
-
-        std::bitset<64> bitset;
-        WriteMultiBitValue(bitset, TestEnum::Value, mask, shift);
-        EmitTestResult(bitset.to_ullong() == 0x0000'0000'0000'0028, "Write multi bit enum mask/shift");
-        EmitTestResult(ReadMultiBitValue<TestEnum>(bitset, mask, shift) == TestEnum::Value, "Read/write multi bit enum round-trip");
-    }
 
     ///////////////////////////////////////////////////////////////////////////
     // Other tests
@@ -235,9 +196,7 @@ namespace UnitTests
 
         MemoryManager::Run();
         Print::Run();
-
-        ReadWriteMultiBitValueTest();
-        ReadWriteMultiBitEnumTest();
+        Utils::Run();
     }
 
     void RunPostStaticDestructors()
