+++ kernel/UnitTests/UtilsTests.h
@@ -1,0 +1,12 @@
+#ifndef KERNEL_UNITTESTS_UTILSTESTS_H
+#define KERNEL_UNITTESTS_UTILSTESTS_H
+
+namespace UnitTests::Utils
+{
+    /**
+     * Run all runtime tests
+     */
+    void Run();
+}
+
+#endif // KERNEL_UNITTESTS_UTILSTESTS_H
